+++ modules/frontend/outputs.tf
@@ -14,6 +14,16 @@
  * limitations under the License.
  */
 
+output "ip_address_internal_managed_http" {
+  description = "The internal/external IP addresses assigned to the HTTP forwarding rules."
+  value       = [for rule in google_compute_global_forwarding_rule.internal_managed_http : rule.ip_address]
+}
+
+output "ip_address_internal_managed_https" {
+  description = "The internal/external IP addresses assigned to the HTTPS forwarding rules."
+  value       = [for rule in google_compute_global_forwarding_rule.internal_managed_https : rule.ip_address]
+}
+
 output "external_ip" {
   description = "The external IPv4 assigned to the global fowarding rule."
   value       = local.address