+++ test/setup/main.tf
@@ -26,6 +26,7 @@ module "project-ci-lb-http" {
   default_service_account     = "keep"
   disable_dependent_services  = false
   disable_services_on_destroy = false
+  deletion_policy             = "DELETE"
 
   activate_apis = [
     "cloudresourcemanager.googleapis.com",
@@ -35,6 +36,7 @@ module "project-ci-lb-http" {
     "run.googleapis.com",
     "iam.googleapis.com",
     "certificatemanager.googleapis.com",
+    "vpcaccess.googleapis.com",
   ]
 }
 
@@ -50,6 +52,7 @@ module "project-ci-lb-http-1" {
   default_service_account     = "keep"
   disable_dependent_services  = false
   disable_services_on_destroy = false
+  deletion_policy             = "DELETE"
 
   activate_apis = [
     "cloudresourcemanager.googleapis.com",
@@ -59,5 +62,6 @@ module "project-ci-lb-http-1" {
     "run.googleapis.com",
     "iam.googleapis.com",
     "certificatemanager.googleapis.com",
+    "vpcaccess.googleapis.com",
   ]
 }